refactor: Prevent cyclic dependencies
This refactors the code to prevent most of the cyclic dependencies.
