+++ src/cli/middlewares/base.ts
@@ -2,13 +2,13 @@ import {initLogger} from "./logger.js";
 import updateNotifier from "./updateNotifier.js";
 
 import type {ArgumentsCamelCase} from "yargs";
-import type {LinterArg} from "../base.ts";
+import type LinterArgs from "../LinterArgs.js";
 /**
  * Base middleware for CLI commands.
  *
  * This middleware should be executed for every CLI command to enable basic features (e.g. logging).
  */
-export default async function (argv: ArgumentsCamelCase<LinterArg>) {
+export default async function (argv: ArgumentsCamelCase<LinterArgs>) {
 	await initLogger(argv);
 	await updateNotifier();
 }
