+++ src/linter/messages.ts
@@ -1,7 +1,5 @@
 // TODO: Migrate to enum instead of Object/Map
 
-import GlobalFix from "./ui5Types/fix/GlobalFix.js";
-
 // Currently, it's done this way to avoid pollution of the test snapshots
 export const RULES = {
 	"async-component-flags": "async-component-flags",
@@ -440,7 +438,7 @@ export const MESSAGE_INFO = {
 		severity: LintMessageSeverity.Error,
 		ruleId: RULES["no-globals"],
 
-		message: ({variableName, namespace}: {variableName: string; namespace: string; fix?: GlobalFix}) =>
+		message: ({variableName, namespace}: {variableName: string; namespace: string}) =>
 			`Access of global variable '${variableName}' (${namespace})`,
 		details: () =>
 			`Do not use global variables to access UI5 modules or APIs. ` +
