+++ src/linter/ui5Types/asyncComponentFlags.ts
@@ -4,7 +4,7 @@ import type SourceFileReporter from "./SourceFileReporter.js";
 import type {JSONSchemaForSAPUI5Namespace, SAPJSONSchemaForWebApplicationManifestFile} from "../../manifest.js";
 import LinterContext from "../LinterContext.js";
 import jsonMap from "json-source-map";
-import type {jsonSourceMapType} from "../manifestJson/ManifestLinter.js";
+import type {jsonSourceMapType} from "../manifestJson/parser.js";
 import {MESSAGE} from "../messages.js";
 import {getPropertyNameText} from "./utils/utils.js";
